refactor(lexer): move `unicode_ws_len` out of `classify` stage (#26484)
Pure refactor. No code changes, just moving code around.

Move `unicode_ws_len` out of `classify` stage. It's used across `carve`, `misc`, `find`, and `regex_div` - but not `classify` (since #26482).

Move it into `find.rs`, which contains all the other primitives reading from the source buffer.
